Blase Data Object

For creation details, see as.BlaseData()

Value

An BlaseData object

Slots

pseudobulk_bins

list of data.frames. Each item is a normalised count matrix representing a bin, where a column is a cell in the bin and each row is a gene.

bins

list. A list of bin names for each timepoint.

genes

list. A list of the genes selected for discriminating timepoints.

Examples

counts <- matrix(rpois(100, lambda = 10), ncol=10, nrow=10)
sce <- SingleCellExperiment::SingleCellExperiment(assays = list(normcounts = counts))
sce$pseudotime = seq_len(10)
as.BlaseData(sce, pseudotime_slot="pseudotime", n_bins=3)
#> Blase Data with:
#> 	bins: c(1, 2, 3)
#> 	selected genes: character(0)