gene_selection_matrix

Generates a plot showing the progression of genes and cells through pseudotime. Plots are of log normalised expression values.

Usage

gene_selection_matrix(x, ...)

# S4 method for Seurat
gene_selection_matrix(
  x,
  waves,
  genes = c(),
  pseudotime_slot = "slingPseudotime_1",
  target_matrix_size = 1000,
  n_cores = 1
)

# S4 method for SingleCellExperiment
gene_selection_matrix(
  x,
  waves,
  genes = c(),
  pseudotime_slot = "slingPseudotime_1",
  target_matrix_size = 1000,
  n_cores = 1
)

Arguments

x

Object to generate the gene selection matrix from

...

Parameters passed to implementations.

waves

Waves dataframe generated by get_waves

genes

The genes to generate matrix for. When an empty list, all genes are used. Defaults to c().

pseudotime_slot

The slot of pseudotime to use for ordering. Defaults to slingPseudotime_1".

target_matrix_size

Matrix size to redimension to. Defaults to 1000.

n_cores

Number of cores for the matrix redimensioning

Value

A plot of the genes and cells ordered by pseudotime