Weng et al. 2024 Haematopoiesis Dataset

This is a SingleCellExperiment object which contains 6000 cells and 1500 highly variable genes. It is used to test BLASE for a Haematopoietic dataset.

Usage

data(processed_simulated_data_SCE, package='blase')

Format

The object contains raw, normalized and log normalized counts, cluster names, and PCA and UMAP coordinates, and Slingshot and TradeSeq results.

Source

BLASE

Details

Exact steps for generating this object are given in our article.